+++ compilers/rust/src/frontend/parser_move.rs
@@ -133,7 +133,7 @@ enum Token {
 
     // Identifiers and literals
     Ident(String),
-    NumberLit(i64),
+    NumberLit(i128),
     StringLit(String),
 
     // Operators
@@ -350,7 +350,7 @@ fn tokenize(source: &str) -> Vec<Token> {
                 .iter()
                 .filter(|c| **c != '_' && **c != 'n')
                 .collect();
-            let val = num_str.parse::<i64>().unwrap_or(0);
+            let val = num_str.parse::<i128>().unwrap_or(0);
             tokens.push(Token::NumberLit(val));
             continue;
         }
